chore: use leak flag to determine whether leak should be check
