+++ backend/internal/jobs/websiteupdate/websiteupdate_test.go
@@ -1,12 +1,26 @@
 package websiteupdate
 
 import (
+	"flag"
+	"os"
 	"testing"
 	"time"
 
 	"github.com/htchan/WebHistory/internal/repository"
 	"github.com/stretchr/testify/assert"
+	"go.uber.org/goleak"
 )
+
+func TestMain(m *testing.M) {
+	leak := flag.Bool("leak", false, "check for memory leaks")
+	flag.Parse()
+
+	if *leak {
+		goleak.VerifyTestMain(m)
+	} else {
+		os.Exit(m.Run())
+	}
+}
 
 func TestSetup(t *testing.T) {
 	t.Parallel()
